RavenDB-20861 Allow Corax.EntryModifications to accept a value and frequency in one shot

+++ src/Corax/Utils/EntryTermsReader.cs
@@ -84,7 +84,7 @@ public unsafe struct EntryTermsReader
     public double Latitude;
     public double Longitude;
     public UnmanagedSpan? StoredField;
-    public int Frequency;
+    public short Frequency;
     public bool HasNumeric;
     public bool IsRaw;
     public bool IsList;
